Tests for and re-runs PMS if previously running
Tested and working (x64).
Upgrades cause PMS to shutdown.
This checks to see if PMS is running or not before installing/upgrading. If PMS is found running, it re-runs PMS after the upgrade. Does not run, just as default, on a new install.
No more finding my PMS not running due to an upgrade and having to pin it to avoid this problem and do manual upgrades. Yay!