(#54) Improves Parameters in Test and Update Scripts
When testing and validating the updates in the scripts, I have found myself to want an autocompleter for the package IDs.

This commit adds an argument completer to complete name (package ID) and forcedpackages.

Also makes some minor changes to improve understandability of the code, and removes some excess whitespace.